+++ WebContent/ShowEmployeeList.html
@@ -4,6 +4,7 @@
 <meta charset="ISO-8859-1">
 <link href="https://fonts.googleapis.com/icon?family=Material+Icons"	rel="stylesheet">
 <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.0/css/bootstrap.min.css">
+<link rel="stylesheet" href="css/show-employee.css">
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
 <script src="https://code.jquery.com/ui/1.12.1/jquery-ui.js"></script>
 <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.0/js/bootstrap.min.js"></script>
@@ -38,43 +39,7 @@ function getEmployees(){
 <body onload="getEmployees()">
       <div class = 'row'>
 	</div>
-	<script src="js/main.js"></script> 
+	<script src="js/ShowEmployee.js"></script> 
 </body>
-<style>
-
-
-div.column {
-  float: left;
-  width: 33.3%;
-  margin-bottom: 16px;
-  padding: 0 8px;
-}
-
-@media screen and (max-width: 650px) {
-  div.column {
-    width: 100%;
-    display: block;
-  }
-}
-
-div.card {
-  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
-}
-
-div.container {
-  padding: 0 16px;
-  font-size: 15px;
-  font-weight:700;
-  
-}
-
-div.container::after, div.row::after {
-  content: "";
-  clear: both;
-  display: table;
-}
-
-
-</style>
 
 </html>
