+++ WebContent/WEB-INF/web.xml
@@ -2,6 +2,6 @@
 <web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://java.sun.com/xml/ns/javaee" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd" id="WebApp_ID" version="3.0">
   <display-name>PlanonEmployeeInformationSystem</display-name>
   <welcome-file-list>
-    <welcome-file>html/index.html</welcome-file>
+    <welcome-file>/index.html</welcome-file>
   </welcome-file-list>
 </web-app>
