--- WebContent/css/styles.css
@@ -1,204 +1,0 @@
-
-* {box-sizing: border-box}
-body {
-	width:100%;
-	 height:100%;
-	 padding:0;
-	font-family: 'Montserrat', sans-serif;
-	line-height:1.6;
-	font-size:17px;
-	 margin:0}
-	 
-/* Slideshow container */
-.slideshow-container {
-   display: block;
-  background: #FFF;
-  overflow: hidden;
-  position: relative;
-  left: 0;
-  right: 0;
-  top: 0;
-  bottom: 0;
-}
-.mySlides {display:none}
-	
-
-.mySlides {
-  display: flex;
-  justify-content: flex-start;
-}
-.slider-content {
-  width: 635px;
-  padding-left: 100px;
-}
-
-.mySlides
-  h4 {font-size: 22px;
-  font-family: &quot;Oswald&quot;;
-  color: white;	
-  font-weight: bold;
-  text-transform: uppercase;
-  line-height: 1.2;
- 
-}
-
-.myslides
-img {
-position: realtive;
-  right: 10%;
-  bottom: 0;
-  height: 100%;  
-}
-
-.button1 {
-  background-color: blue;
-  width: 220px;
-  height: 50px;
-  border-radius: 30px;
-  border: none;
-  font-family: Montserrat;
-  font-size: 20px;
-  font-weight: 100;
-  color: #fff;
-  text-align: left;
-  padding-left: 35px;
-  position: relative;
-  cursor: pointer;
-  transition: all .2s;
-}
-
-
-.button1:hover {
-  box-shadow: 0px 0px 60px -17px rgba(51,51,51,1);
-}
- 
-
-
-
-/* Next & previous buttons */
-.prev, .next {
-  cursor: pointer;
-  position: absolute;
-  top: 50%;
-  width: auto;
-  margin-top: -22px;
-  padding: 16px;
-  color: white;
-  font-weight: bold;
-  font-size: 18px;
-  transition: 0.6s ease;
-  border-radius: 0 3px 3px 0;
-  user-select: none;
-}
-
-/* Position the "next button" to the right */
-.next {
-  right: 0;
-  border-radius: 3px 0 0 3px;
-}
-
-/* On hover, add a black background color with a little bit see-through */
-.prev:hover, .next:hover {
-  background-color: rgba(0,0,0,0.8);
-}
-
-/* Caption text */
-.slider-content {
-  color: #f2f2f2;
-  font-size: 15px;
-  padding: 8px 12px;
-  position: absolute;
-  bottom: 8px;
-  width: 100%;
-  text-align: center;
-}
-
-/* Number text (1/3 etc) */
-.numbertext {
-  color: #f2f2f2;
-  font-size: 12px;
-  padding: 8px 12px;
-  position: absolute;
-  top: 0;
-}
-
-/* The dots/bullets/indicators */
-.dot {
-  cursor: pointer;
-  height: 15px;
-  width: 15px;
-  margin: 0 2px;
-  background-color: #bbb;
-  border-radius: 50%;
-  display: inline-block;
-  transition: background-color 0.6s ease;
-}
-
-.active, .dot:hover {
-  background-color: #717171;
-}
-
-/* Fading animation */
-.fade {
-  -webkit-animation-name: fade;
-  -webkit-animation-duration: 1.5s;
-  animation-name: fade;
-  animation-duration: 1.5s;
-}
-
-@-webkit-keyframes fade {
-  from {opacity: .4} 
-  to {opacity: 1}
-}
-
-@keyframes fade {
-  from {opacity: .4} 
-  to {opacity: 1}
-}
-
-.button {
-	padding:1em 2em;
-	color:#fff; 
-	border:0;
-}
-
-.editform{
-	 cursor: pointer;
-	  border: none;
-	  background-color: coral;
-	  font-size:20px;
-	  float:right;
-	  margin:10px;
-}
-
-
-#closeBtn {
-	color:#ccc;
-	float:right;
-	font-size:30px;
-	display:inline;
-}
-
-#closeBtn1 {
-	color:#ccc;
-	float:right;
-	font-size:30px;
-}
-
-#closeBtn:hover, #closeBtn:focus {
-color:#000;
-text-decoration:none;
-cursor:pointer;
-}
-
-@keyframes modalopen{
-	from{ opacity : 0}
-	to {opacity:1}
-}
-
-.message{
-	text-color:black;
-}
-.ui-dialog .ui-dialog-titlebar .ui-dialog-titlebar-close {
-    display: none;
-}
