--- WebContent/css/styles1.css
@@ -1,40 +1,0 @@
-input:not([type=radio]), input:focus, select, textarea {
-  width: 100%;
-  padding: 12px;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  box-sizing: border-box;
-  resize: vertical;
-  outline:none;
-  overflow: hidden;
-  word-wrap: normal;
-  white-space: normal;
-}
-
-label {
-  padding: 12px 12px 12px 0;
-  display: inline-block;
-}
-
-/* Style the submit button */
-input[type=submit] {
-  background-color: #4CAF50;
-  color: white;
-  padding: 12px 20px;
-  border: none;
-  cursor: pointer;
-}
-
-input[type=reset] {
-  background-color: #CD5C5C;
-  color: white;
-  padding: 12px 20px;
-  border: none;
-  cursor: pointer;
-  width:7em
-}
-
-.myClass{
-background: #EBCCCC;
-  font-size: 1em;
-}
