--- src/com/Welcome.java
@@ -1,5 +1,0 @@
-package com;
-
-public class Welcome {
-
-}
