+++ src/com/planon/ped/common/IEmployeeService.java
@@ -1,6 +1,7 @@
 package com.planon.ped.common;
 
 import java.sql.SQLException;
+import java.util.List;
 
 import com.planon.ped.dto.EmployeeData;
 
@@ -11,4 +12,7 @@ public interface IEmployeeService {
 	void updateEmployeeDetails(EmployeeData aEmployee) throws SQLException;
 
 	void deleteEmployee(int aEmployeeCode) throws SQLException;
+	
+	List<EmployeeData> getListOfAllEmployees() throws SQLException;
+	
 }
