--- src/com/planon/ped/dto/AbstractData.java
@@ -1,33 +1,0 @@
-package com.planon.ped.dto;
-
-
-abstract public class AbstractData   {
-	private int code;
-	private String name;
-	private String location;
-
-	public int getCode() {
-		return code;
-	}
-
-	public void setCode(int code) {
-		this.code = code;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public String getLocation() {
-		return location;
-	}
-
-	public void setLocation(String location) {
-		this.location = location;
-	}
-
-}
