+++ src/com/planon/resources/scripts.txt
@@ -1,19 +1,6 @@
-CREATE TABLE [dbo].[TEAM](
-	[CODE] [int] IDENTITY(1,1) NOT NULL,
-	[NAME] [varchar](50) NULL,
-	[DESCRIPTION] [varchar](255) NULL,
-	[TYPE] [varchar](25) NULL,
-	[LOCATION] [varchar](255) NULL,
-PRIMARY KEY CLUSTERED 
-(
-	[CODE] ASC
-)WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON) ON [PRIMARY]
-) ON [PRIMARY]
-
 CREATE TABLE [dbo].[EMPLOYEE](
 	[CODE] [int] IDENTITY(1,1) NOT NULL,
 	[NAME] [varchar](50) NULL,
-	[TEAM_REF] [int] NULL,
 	[EMAIL] [varchar](100) NULL,
 	[PHONE] [varchar](50) NULL,
 	[GENDER] [varchar](1) NULL,
@@ -26,7 +13,3 @@ PRIMARY KEY CLUSTERED
 )WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON) ON [PRIMARY]
 ) ON [PRIMARY]
 
-
-
-ALTER TABLE [dbo].[EMPLOYEE]  WITH CHECK ADD FOREIGN KEY([TEAM_REF])
-REFERENCES [dbo].[TEAM] ([CODE])
